Rename WgpuAutoGui to EventTarget and move to events module
