+++ wgpu/gui/glfw.py
@@ -17,7 +17,8 @@ import traceback
 
 import glfw
 
-from .base import WgpuCanvasBase, WgpuAutoGui
+from .base import WgpuCanvasBase
+from .events import EventTarget
 
 
 # Make sure that glfw is new enough
@@ -122,7 +123,7 @@ KEY_MAP_MOD = {
 }
 
 
-class GlfwWgpuCanvas(WgpuAutoGui, WgpuCanvasBase):
+class GlfwWgpuCanvas(EventTarget, WgpuCanvasBase):
     """A glfw window providing a wgpu canvas."""
 
     # See https://www.glfw.org/docs/latest/group__window.html
