+++ wgpu/gui/offscreen.py
@@ -3,10 +3,10 @@ import asyncio
 import numpy as np
 
 from ._offscreen import WgpuOffscreenCanvas
-from .base import WgpuAutoGui
+from .events import EventTarget
 
 
-class WgpuManualOffscreenCanvas(WgpuAutoGui, WgpuOffscreenCanvas):
+class WgpuManualOffscreenCanvas(EventTarget, WgpuOffscreenCanvas):
     """An offscreen canvas intended for manual use. Call the ``.draw()``
     method to perform a draw and get the result.
     """
